gate_yolo_mvp/requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8.0
PyYAML>=6.0
numpy>=1.24
fastapi>=0.115.0
uvicorn>=0.30.0

Flask>=3.0.0
Pillow>=10.0.0
